Friendbot, comments: move horizon/httpx package back to horizon/internal/httpx
