+++ services/horizon/internal/resource/operations/base.go
@@ -3,8 +3,8 @@ package operations
 import (
 	"fmt"
 
-	"github.com/stellar/go/services/horizon/httpx"
 	"github.com/stellar/go/services/horizon/internal/db2/history"
+	"github.com/stellar/go/services/horizon/internal/httpx"
 	"github.com/stellar/go/services/horizon/internal/render/hal"
 	"golang.org/x/net/context"
 )
