+++ services/horizon/internal/resource/transaction.go
@@ -6,8 +6,8 @@ import (
 	"time"
 
 	"github.com/guregu/null"
-	"github.com/stellar/go/services/horizon/httpx"
 	"github.com/stellar/go/services/horizon/internal/db2/history"
+	"github.com/stellar/go/services/horizon/internal/httpx"
 	"github.com/stellar/go/services/horizon/internal/render/hal"
 	"golang.org/x/net/context"
 )
