+++ services/horizon/internal/resource/transaction_success.go
@@ -1,9 +1,9 @@
 package resource
 
 import (
-	"github.com/stellar/go/services/horizon/httpx"
-	"github.com/stellar/go/services/horizon/internal/txsub"
+	"github.com/stellar/go/services/horizon/internal/httpx"
 	"github.com/stellar/go/services/horizon/internal/render/hal"
+	"github.com/stellar/go/services/horizon/internal/txsub"
 	"golang.org/x/net/context"
 )
